fix: #75 extract GradientBackground and dashboard components
- Create shared GradientBackground component with 'simple' and 'animated' variants
- Extract UpdateBanner, DashboardGreeting, and DashboardChatList from dashboard.tsx
