+++ app/download.tsx
@@ -1,11 +1,11 @@
 import { FlappyBird } from "@/components/flappy-bird";
+import { GradientBackground } from "@/components/gradient-background";
 import { Button } from "@/components/ui/button";
 import { ModeToggle } from "@/components/ui/mode-toggle";
 import { Progress } from "@/components/ui/progress";
 import { Text } from "@/components/ui/text";
 import { View } from "@/components/ui/view";
 import { useColor } from "@/hooks/useColor";
-import { useColorScheme } from "@/hooks/useColorScheme";
 import { DEFAULT_AI_CHAT_MODEL } from "@/src/actions/ai/constants";
 import {
 	pauseDownload,
@@ -14,18 +14,15 @@ import {
 import { fetchLatestRecommendedModel } from "@/src/actions/ai/model-config";
 import { completeOnboarding } from "@/src/actions/settings";
 import { mainStore } from "@/src/stores/main/main-store";
-import { ImageBackground } from "expo-image";
 import { useRouter } from "expo-router";
 import { useEffect, useState } from "react";
-import { ActivityIndicator, Dimensions, Modal } from "react-native";
+import { ActivityIndicator, Modal } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
-import { Defs, RadialGradient, Rect, Stop, Svg } from "react-native-svg";
 import { useValue } from "tinybase/ui-react";
 import type { WhisperLLMCard } from "whisper-llm-cards";
 
 export default function Download() {
 	const router = useRouter();
-	const scheme = useColorScheme();
 	const backgroundColor = useColor("background");
 	const primaryForegroundColor = useColor("primaryForeground");
 
@@ -156,90 +153,7 @@ export default function Download() {
 
 	return (
 		<View style={{ flex: 1 }}>
-			{/* Background gradient - same as onboarding */}
-			<View
-				style={{
-					position: "absolute",
-					top: 0,
-					left: 0,
-					width: "100%",
-					height: "100%",
-					flex: 1,
-					display: "flex",
-				}}
-			>
-				<Svg
-					key={scheme}
-					style={[
-						{
-							flex: 1,
-							position: "absolute",
-							top: 0,
-							left: 0,
-							width: "100%",
-							height: Dimensions.get("window").height,
-						},
-					]}
-					viewBox={`0 0 1 ${Dimensions.get("window").height / Dimensions.get("window").width}`}
-				>
-					<Defs>
-						<RadialGradient
-							id="radialGradient"
-							gradientUnits="objectBoundingBox"
-							cx={0.5}
-							cy={0.5}
-							r={0.75}
-						>
-							<Stop
-								offset="0"
-								stopColor={"#ff9e37ff"}
-								stopOpacity={scheme !== "dark" ? 0.9 : 0.95}
-							/>
-							<Stop
-								offset="0.275"
-								stopColor={"#ff5b91ff"}
-								stopOpacity={scheme !== "dark" ? 0.9 : 0.9}
-							/>
-							<Stop
-								offset="0.3"
-								stopColor={"#ff95ffff"}
-								stopOpacity={scheme !== "dark" ? 0.8 : 0.6}
-							/>
-							<Stop
-								offset="0.325"
-								stopColor={"#69b7ffff"}
-								stopOpacity={scheme !== "dark" ? 0.7 : 0.6}
-							/>
-							<Stop
-								offset="0.65"
-								stopColor={"#0017c2ff"}
-								stopOpacity={scheme !== "dark" ? 0 : 0}
-							/>
-							<Stop offset="0.85" stopColor={"transparent"} stopOpacity={0} />
-						</RadialGradient>
-					</Defs>
-					<Rect
-						x={-1.5}
-						y={0.125}
-						width="4"
-						height="4"
-						fill="url(#radialGradient)"
-					/>
-				</Svg>
-
-				<ImageBackground
-					source={
-						scheme === "dark"
-							? require(`../assets/images/grain-dark.png`)
-							: require(`../assets/images/grain.png`)
-					}
-					style={{
-						flex: 1,
-						opacity: scheme === "dark" ? 0.3 : 0.25,
-						backgroundColor: backgroundColor,
-					}}
-				/>
-			</View>
+			<GradientBackground variant="simple" />
 
 			<SafeAreaView style={{ flex: 1 }}>
 				{/* Mode toggle in top right */}
